+++ src/main/java/dev/juanleon/supermarket_inventory/products/domain/services/get/IGetProductsServices.java
@@ -4,6 +4,7 @@ import dev.juanleon.supermarket_inventory.common.utils.dto.PagedResponse;
 import dev.juanleon.supermarket_inventory.common.utils.dto.PaginationRequest;
 import dev.juanleon.supermarket_inventory.products.domain.models.ProductModel;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -15,8 +16,8 @@ public interface IGetProductsServices {
     PagedResponse<ProductModel> getByName(PaginationRequest paginationRequest, String name);
     PagedResponse<ProductModel> getByActive(PaginationRequest paginationRequest, Boolean active);
     PagedResponse<ProductModel> getByStock(PaginationRequest paginationRequest, Integer stock);
-    PagedResponse<ProductModel> getByPriceSale(PaginationRequest paginationRequest, Double priceSale);
-    PagedResponse<ProductModel> getByPricePurchase(PaginationRequest paginationRequest, Double pricePurchase);
+    PagedResponse<ProductModel> getByPriceSale(PaginationRequest paginationRequest, BigDecimal priceSale);
+    PagedResponse<ProductModel> getByPricePurchase(PaginationRequest paginationRequest, BigDecimal pricePurchase);
     PagedResponse<ProductModel> getByUnitMeasurement(PaginationRequest paginationRequest, String unitMeasurement);
     PagedResponse<ProductModel> getByCreatedAt(PaginationRequest paginationRequest, LocalDate createdAt);
     PagedResponse<ProductModel> getByUpdatedAt(PaginationRequest paginationRequest, LocalDate updatedAt);
